refactor: syntax code go (#29)
