+++ advanced/goroutines/yield.go
@@ -1,8 +1,29 @@
-package main
+package goroutines
 
 import "fmt"
 
-func fib(c chan int, n int) {
+func fib(n int) chan int {
+	C := make(chan int)
+
+	go func() {
+		a, b := 0, 1
+		for i := 0; i < n; i++ {
+			a, b = b, a+b
+			C <- a
+		}
+		close(C)
+	}()
+
+	return C
+}
+
+func getFib() {
+	for x := range fib(10) {
+		fmt.Println(x)
+	}
+}
+
+func fibBest(c chan int, n int) {
 	a, b := 0, 1
 	for i := 0; i < n; i++ {
 		a, b = b, a+b
@@ -11,9 +32,9 @@ func fib(c chan int, n int) {
 	close(c)
 }
 
-func main() {
+func getFibBest() {
 	c := make(chan int)
-	go fib(c, 10)
+	go fibBest(c, 10)
 
 	for x := range c {
 		fmt.Println(x)
