+++ beginner/pointers/example1.go
@@ -2,5 +2,6 @@ package beginner
 
 func newReleaseArtist(a Artist) int {
 	a.Songs++
+
 	return a.Songs
 }
