+++ beginner/pointers/example2.go
@@ -2,5 +2,6 @@ package beginner
 
 func newRelease(a *Artist) int {
 	a.Songs++
+
 	return a.Songs
 }
