+++ beginner/strings/example1.go
@@ -1,23 +1,13 @@
-package main
+package beginner
 
-import (
-	"fmt"
-)
+func getGreetingName() string {
+	s1 := "Hola, me llamo Luis Mayta"
 
-func example1() {
-	var s1 string = "Hola, me llamo Luis Mayta"
-	fmt.Println(s1)
-	fmt.Println("¿Cómo has dicho que te llamas?")
-	fmt.Println(s1[15:])
+	return s1[15:]
 }
 
-func example2() {
-	var s1 string = "Hola, me llamo Íñigo Montoya"
-	s1[3] = 'X'
-	fmt.Println(s1)
-}
+func getGreeting() string {
+	s1 := "Hola, me llamo LM"
 
-func main() {
-	example1()
-	// example2()
+	return s1
 }
