+++ beginner/variables/base.go
