+++ beginner/variables/functions.go
