+++ beginner/variables/single.go
@@ -4,5 +4,6 @@ import "fmt"
 
 func getSingleVar() string {
 	name, location := "itachi", "konoha"
+
 	return fmt.Sprintf("%s of %s", name, location)
 }
