+++ beginner/variables/vars_replace.go
@@ -8,5 +8,6 @@ func getMessage() string {
 	name := "itachi uchiha"
 
 	message := fmt.Sprintf("Hello %s", name)
+
 	return message
 }
